Update: Create basic react tests
