+++ .babelrc
@@ -1,6 +1,7 @@
 {
     "presets": [
         "es2015",
-        "stage-1"
+        "stage-1",
+        "react"
     ]
 }
