+++ package.json
@@ -34,7 +34,9 @@
     "babel-istanbul-loader": "0.0.1",
     "babel-loader": "^6.2.1",
     "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
     "babel-preset-stage-1": "^6.3.13",
+    "es5-shim": "^4.4.1",
     "eslint": "^1.10.3",
     "eslint-config-airbnb": "^3.1.0",
     "eslint-plugin-react": "^3.15.0",
@@ -43,12 +45,16 @@
     "karma-coverage": "^0.5.3",
     "karma-coveralls": "^1.1.2",
     "karma-jasmine": "^0.3.6",
+    "karma-jasmine-html-reporter-livereload": "^1.0.0",
     "karma-phantomjs-launcher": "^0.2.3",
     "karma-sourcemap-loader": "^0.3.6",
     "karma-webpack": "^1.7.0",
     "lodash": "^3.10.1",
     "moment": "^2.11.1",
     "phantomjs": "^1.9.19",
+    "react": "^0.14.6",
+    "react-addons-test-utils": "^0.14.6",
+    "react-dom": "^0.14.6",
     "release-it": "^2.3.1",
     "webpack": "^1.12.11"
   }
